Added Timelock to Sonic deploy script
(cherry picked from commit 2c3e2e4e90de17dc362e3a2c273dc0cfa8b652fd)
