+++ contracts/hardhat.config.js
@@ -310,6 +310,8 @@ module.exports = {
       localhost:
         process.env.FORK_NETWORK_NAME == "base"
           ? addresses.base.timelock
+          : process.env.FORK_NETWORK_NAME == "sonic"
+          ? addresses.sonic.timelock
           : process.env.FORK_NETWORK_NAME == "mainnet" ||
             (!process.env.FORK_NETWORK_NAME && process.env.FORK == "true")
           ? MAINNET_TIMELOCK
@@ -317,6 +319,8 @@ module.exports = {
       hardhat:
         process.env.FORK_NETWORK_NAME == "base"
           ? addresses.base.timelock
+          : process.env.FORK_NETWORK_NAME == "sonic"
+          ? addresses.sonic.timelock
           : process.env.FORK_NETWORK_NAME == "mainnet" ||
             (!process.env.FORK_NETWORK_NAME && process.env.FORK == "true")
           ? MAINNET_TIMELOCK
