+++ contracts/utils/addresses.js
@@ -350,7 +350,7 @@ addresses.sonic.validatorRegistrator =
 addresses.sonic.admin = "0xAdDEA7933Db7d83855786EB43a238111C69B00b6";
 // 2/8 Multisig - formally known as the Strategist
 addresses.sonic.guardian = "0x63cdd3072F25664eeC6FAEFf6dAeB668Ea4de94a";
-// addresses.sonic.timelock = "";
+addresses.sonic.timelock = "0x31a91336414d3B955E494E7d485a6B06b55FC8fB";
 
 // Holesky
 addresses.holesky.WETH = "0x94373a4919B3240D86eA41593D5eBa789FEF3848";
